SES-2 move existed components in their own files also export them as default
