+++ src/App.jsx
@@ -1,0 +1,17 @@
+import Header from "./components/Header"
+import Content from "./components/Content"
+import Footer from "./components/Footer"
+
+import "./App.css";
+
+function App() {
+  return (
+    <div className="bg-primary text-white">
+      <Header />
+      <Content />
+      <Footer />
+    </div>
+  );
+}
+
+export default App;
