+++ src/components/Content.jsx
@@ -1,0 +1,11 @@
+import Products from './Products';
+
+const Content = () => {
+    return (
+        <main className="flex flex-col justify-between">
+            <Products />
+        </main>
+    );
+}
+
+export default Content;
